+++ import_plugin/create_func.sql
@@ -1,6 +1,6 @@
 --
 --    Ophidia Primitives
---    Copyright (C) 2012-2022 CMCC Foundation
+--    Copyright (C) 2012-2025 CMCC Foundation
 --
 --    This program is free software: you can redistribute it and/or modify
 --    it under the terms of the GNU General Public License as published by
@@ -100,6 +100,7 @@ DROP FUNCTION IF EXISTS oph_aggregate_stats_partial;
 DROP FUNCTION IF EXISTS oph_aggregate_stats_final;
 DROP FUNCTION IF EXISTS oph_mask_array;
 DROP FUNCTION IF EXISTS oph_interlace;
+DROP FUNCTION IF EXISTS oph_interlace2;
 DROP FUNCTION IF EXISTS oph_append;
 DROP FUNCTION IF EXISTS oph_quantize;
 DROP FUNCTION IF EXISTS oph_gsl_spline;
@@ -117,6 +118,7 @@ DROP FUNCTION IF EXISTS oph_sequence;
 DROP FUNCTION IF EXISTS oph_predicate;
 DROP FUNCTION IF EXISTS oph_predicate2;
 DROP FUNCTION IF EXISTS oph_predicate3;
+DROP FUNCTION IF EXISTS oph_matheval;
 
 DELIMITER //
 CREATE PROCEDURE mysql.oph_drill_down(IN table_in VARCHAR(100), IN outer_size INT, IN inner_size INT, IN oph_type VARCHAR(30), IN table_out VARCHAR(100), IN compressed INT)
@@ -239,6 +241,7 @@ CREATE AGGREGATE FUNCTION oph_aggregate_stats_partial RETURNS STRING SONAME 'lib
 CREATE AGGREGATE FUNCTION oph_aggregate_stats_final RETURNS STRING SONAME 'liboph_aggregate_stats_final.so';
 CREATE FUNCTION oph_mask_array RETURNS STRING SONAME 'liboph_mask_array.so';
 CREATE FUNCTION oph_interlace RETURNS STRING SONAME 'liboph_interlace.so';
+CREATE FUNCTION oph_interlace2 RETURNS STRING SONAME 'liboph_interlace2.so';
 CREATE FUNCTION oph_append RETURNS STRING SONAME 'liboph_append.so';
 CREATE FUNCTION oph_quantize RETURNS STRING SONAME 'liboph_quantize.so';
 CREATE FUNCTION oph_gsl_spline RETURNS STRING SONAME 'liboph_gsl_spline.so';
@@ -256,6 +259,7 @@ CREATE FUNCTION oph_sequence RETURNS STRING SONAME 'liboph_sequence.so';
 CREATE FUNCTION oph_predicate RETURNS STRING SONAME 'liboph_predicate.so';
 CREATE FUNCTION oph_predicate2 RETURNS STRING SONAME 'liboph_predicate2.so';
 CREATE FUNCTION oph_predicate3 RETURNS STRING SONAME 'liboph_predicate3.so';
+CREATE FUNCTION oph_matheval RETURNS STRING SONAME 'liboph_matheval.so';
 
 GRANT EXECUTE ON PROCEDURE `mysql`.`oph_subset` TO `%`;
 GRANT EXECUTE ON PROCEDURE `mysql`.`oph_drill_down` TO `%`;