+++ src/oph_matheval.h
@@ -22,8 +22,7 @@
 #include <string.h>
 #include <ctype.h>
 
-/* Core function header */
-#include "oph_core.h"
+#include "oph_core_matheval.h"
 
 /* MySQL headers  */
 #include <mysql.h>		// It contains UDF-related symbols and data structures
@@ -31,14 +30,22 @@
 typedef bool my_bool;
 #endif
 
+typedef struct {
+	void *f[2];		// measure and expressions
+	char is_index;		// is_index
+	unsigned long length;	// size in bytes
+	oph_type result_type;
+	size_t result_elemsize;
+} oph_matheval_param;
+
 /*------------------------------------------------------------------|
 |		Functions' declarations (BEGIN)			    |
 |------------------------------------------------------------------*/
 
 /* These must be right or mysqld will not find the symbol! */
-my_bool oph_concat2_init(UDF_INIT * initid, UDF_ARGS * args, char *message);
-void oph_concat2_deinit(UDF_INIT * initid);
-char *oph_concat2(UDF_INIT * initid, UDF_ARGS * args, char *result, unsigned long *length, char *is_null, char *error);
+my_bool oph_matheval_init(UDF_INIT * initid, UDF_ARGS * args, char *message);
+void oph_matheval_deinit(UDF_INIT * initid);
+char *oph_matheval(UDF_INIT * initid, UDF_ARGS * args, char *result, unsigned long *length, char *is_null, char *error);
 
 /*------------------------------------------------------------------|
 |               Functions' declarations (END)                       |